SUMMARY: HUND in progress
     BUGZILLA BUG #(s):  4517

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  N/A

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  N/A

     3)  Did you modify or add any source or build system files?  yes
     3a) Did you rebuild the libraries and executables?  yes
     3b) Did you rerun the test suite on the packages that you changed and
         all packages that depend on the modified packages?  not tested
     3c) Did you change the required or optional dependencies for any
         package?  no


     DESCRIPTION OF CHANGES:

     Add a cmake flag to do 2D partitioning the HUND way.

(checklist version 2009-07-14)

+++ packages/zoltan/src/phg/phg_build.c
@@ -840,7 +840,7 @@ long int index;
     if (map_num < 0) goto End;
 
     for (i=0; i < zhg->nObj; i++){
-      indexptr = (int *)(long)(i+1);
+      indexptr = (int *)(i+1);
       ierr = Zoltan_Map_Add(zz, map_num, myObjGNO + i, indexptr);
       if (ierr != ZOLTAN_OK) goto End;
     }
